backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
google-genai>=1.0.0
Pillow>=10.0.0
aiofiles>=23.0.0
